fbdev: mdss: Use smaller data type to store soft limits

+++ drivers/video/fbdev/msm/mdss_dsi_panel.c
@@ -58,8 +58,8 @@ void mdss_panel_reset_skip_enable(bool enable)
 	mdss_panel_reset_skip = enable;
 }
 
-static unsigned int bl_level_soft_limit = 16;
-module_param(bl_level_soft_limit, uint, 0644);
+static unsigned short bl_level_soft_limit = 16;
+module_param(bl_level_soft_limit, ushort, 0644);
 
 void mdss_dsi_panel_pwm_cfg(struct mdss_dsi_ctrl_pdata *ctrl)
 {
